Move purchase success display to useDisplayPurchaseSuccess
